(FE) [rename] 레이아웃 파일 이름 오타 수정
